controle_validades/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.0
django4-background-tasks==1.2.9
mysqlclient==2.2.1
pillow==10.2.0
psycopg2==2.9.9
python-dotenv==1.0.0
python-memcached==1.62
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
